refactor(frontend): move user provider into users presentation layer
