+++ frontend/src/modules/users/presentation/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./user-provider";
